PCT_epithelial model with one buffer pair NH3
the simulation goes to steady-state and system sounds reasonable. there are  8 solutes , here is the list of the solutes available in the model.
VARS = [ VE, PE, CE_NA, CE_K,
         CE_CL, CE_HCO3, CE_UREA, CE_NH3, CE_NH4, CE_GLUC,
         VI, IMP, CI_NA, CI_K, CI_CL, CI_HCO3, CI_UREA, CI_NH3, CI_NH4,  CI_GLUC,   VM ].
Here is the changes for CM_Na
CMNA = [0.100, 0.11,  0.12, 0.13, 0.14]

and here is the time step information in regard into the simulation's results.